refactor: 불필요한 코드 리팩터링
