+++ doochul/src/main/java/org/doochul/application/ProductService.java
@@ -1,10 +1,7 @@
 package org.doochul.application;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.doochul.domain.product.Product;
 import org.doochul.domain.product.ProductRepository;
-import org.doochul.ui.dto.ProductResponse;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,8 +10,4 @@ public class ProductService {
 
     private final ProductRepository productRepository;
 
-    public List<ProductResponse> findMemberShipsById(Long userId) {
-        List<Product> products = productRepository.findByUserId(userId);
-        return ProductResponse.fromList(products);
-    }
 }
