+++ doochul/src/main/java/org/doochul/ui/dto/KakaoAccount.java
@@ -1,0 +1,8 @@
+package org.doochul.ui.dto;
+
+import lombok.Data;
+
+@Data
+public class KakaoAccount {
+    private Profile profile;
+}
