+++ doochul/src/main/java/org/doochul/ui/dto/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package org.doochul.domain.oauth.kakao;
+package org.doochul.ui.dto;
 
 import lombok.Data;
